fix(shikomi-infra): add AceFlags==0 check, split windows/ module, add windows CI
修正1 (服部平次): verify_dacl_owner_only の AceType チェック直後に AceFlags == 0 検証を追加。
継承フラグ (CONTAINER_INHERIT_ACE / OBJECT_INHERIT_ACE) を持つ ACE を拒否する。
actual ラベル: "ace_count: 1 (expected 1); ace[0]: ace_flags=0x{XX} (expected 0x00, no inheritance flags)"

修正2 (ペガサス): windows.rs (972行) を 500行制限に適合するよう分割。
  permission/windows/ ディレクトリを新設:
    mod.rs (47行): 定数 + モジュール宣言 + pub(super) 再エクスポート
    guard.rs (93行): SecurityDescriptorGuard / LocalFreeAclGuard / SidStringGuard
    helper.rs (256行): sid_to_string / path_to_wide / fetch_* / build_* / apply_*
    verify.rs (186行): verify_dacl_owner_only / enumerate_aces
    ensure.rs (85行): ensure_dir / ensure_file / verify_dir / verify_file
    tests.rs (363行): TC-U17〜U25
  TC-I24〜I27 を tests/integration_windows.rs に分離 (windows-only binary)。
  SetNamedSecurityInfoW の引数を as_ptr() → as_mut_ptr() に修正 (PWSTR = *mut u16)。

修正3 (まこちゃん): .github/workflows/windows.yml を追加。
  windows-latest ランナーで cargo test -p shikomi-infra を実行。
