+++ .github/workflows/windows.yml
@@ -1,0 +1,23 @@
+name: Test Infra (Windows)
+
+on:
+  pull_request:
+    types: [opened, synchronize, reopened]
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  test-infra-windows:
+    name: test-infra-windows
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: dtolnay/rust-toolchain@stable
+
+      - uses: Swatinem/rust-cache@v2
+
+      - name: cargo test (shikomi-infra)
+        run: cargo test -p shikomi-infra
